refactor(schemagen): run it from cmd/, in the shape of the linter
schemagen sat under tools/ as a flag-parsing main package, which meant
the repository had two different ideas of what a command looks like. It
now follows the one the linter already established: a thin entry point in
cmd/schemagen that only wires streams and exit codes, and the command
itself in pkg/cmd/schemagen with Options, RegisterFlags, Prepare and Run.

Being a package rather than a main is what makes it testable, so the
generator has tests for the first time. They plant source archives in the
cache directory under the name a download would have been saved as, which
exercises the whole pipeline -- harvest, Config-struct fields, overlays,
per-distribution split, index -- without reaching the network.

The flags keep their names but take cobra's double dash, so callers move
from -version to --version. Progress goes through the command's own
writer instead of os.Stdout, and exit codes are exported the way the
linter exports them: 0 written, 1 a harvest that failed, 2 an invocation
that could not run.

Two things are new rather than moved. --formats now refuses a name it
cannot write: Schema.Write treats anything that is not JSON as YAML, so
"toml" used to write YAML under a file name the registry never reads
back. And a stray argument -- "schemagen v0.157.0" for "--version
v0.157.0" -- is now reported as the usage error it is, with the usage
text, rather than as a failed harvest.
