+++ .github/workflows/schemas.yaml
@@ -34,8 +34,8 @@ jobs:
           go-version: 1.25
       - name: Generate
         run: |
-          go run ./tools/schemagen \
-            -version '${{ inputs.versions }}' -out schemas-repo
+          go run ./cmd/schemagen \
+            --version '${{ inputs.versions }}' --out schemas-repo
       - name: Check the generated schemas lint
         run: |
           jq -r '.distributions | to_entries[] | .key as $d | .value[] | "\($d) \(.)"' \
@@ -56,9 +56,9 @@ jobs:
           commit-message: |
             Add schemas for ${{ inputs.versions }}
 
-            Generated by tools/schemagen in otelcol-config-lint.
+            Generated by cmd/schemagen in otelcol-config-lint.
           body: |
-            Generated by [`tools/schemagen`](https://github.com/minuk-dev/otelcol-config-lint/tree/main/tools/schemagen)
+            Generated by [`cmd/schemagen`](https://github.com/minuk-dev/otelcol-config-lint/tree/main/cmd/schemagen)
             from run ${{ github.run_id }}, for `${{ inputs.versions }}`.
 
             Every generated file was linted against `testdata/valid` before this
