+++ pkg/cmd/schemagen/fields.go
@@ -1,4 +1,4 @@
-package main
+package schemagen
 
 import (
 	"path"
